+++ src/Components/ClipCard/index.ts
@@ -1,0 +1,3 @@
+import ClipCard from "./ClipCard"
+
+export default ClipCard
