+++ src/Components/CreateForm/index.ts
@@ -1,0 +1,3 @@
+import CreateForm from "./CreateForm"
+
+export default CreateForm
