+++ src/Components/CreateSearch/index.ts
@@ -1,0 +1,3 @@
+import CreateSearch from "./CreateSearchContainer"
+
+export default CreateSearch
