+++ src/Components/PostCreateHeader/index.ts
@@ -1,0 +1,3 @@
+import PostCreateHeader from "./PostCreateHeader"
+
+export default PostCreateHeader
