+++ src/Routes/CreatePost/CreatePostPresenter.tsx
@@ -2,8 +2,7 @@ import React from "react";
 import PostCreateHeader from "../../Components/PostCreateHeader";
 import styled from "styled-components";
 import ClipList from "../../Components/ClipList";
-import PostForm from "../../Components/PostForm";
-import PostView from "../../Components/PostView";
+import CreateForm from "../../Components/CreateForm";
 
 interface IProps {
   searchedResults: object[];
@@ -60,17 +59,8 @@ export default class CreatePostPresenter extends React.Component<
         );
       case "form":
         return (
-          <PostForm
+          <CreateForm
             selectedResult={this.state.selectedResult}
-          />
-        );
-      case "view":
-        return (
-          <PostView
-            created={{
-              selectedResult: this.state.selectedResult,
-              post: this.state.post
-            }}
           />
         );
       default:
